Solved odd number elements
After watching a video on youtube on class
selectors. I started experimenting by changing
the class name from Number-1 Number-2 and
Number-3 to odd-number.

I created 2 classes in index.html and style.css
which are odd-number and Number-3 and followed the
requirements which is to make all of the elements
to have background color of light pinkish with
Verdana and sans-serif font as fallback.

Additionally Number 3 needs to be 24px in font
size